Test auto-deployment to the PPRD branch
We kick off an Ansible Tower build for merges to the "pprd" branch.  This requires that three secrets be set up as environment variables in Travis:

AWX_USER: Ansible Tower user
AWX_PASS: password for above user
AWX_TEMPLATE_URL: Ansible Tower endpoint URL to launch the appropriate template for this deploy.

To be able to test this currently we have to nullify the actual testing stage because, as of now, the test suite has failing tests.