Trigger Ansible Tower deploy for "pprd" branch
Add a "deploy" stage to the Travis CI build that triggers a launch of the template associated with the "pprd" branch when merges are made to "pprd". This requires that three secrets be set up as environment variables in the Travis settings for this repository:

AWX_USER: Ansible Tower user
AWX_PASS: password for above user
AWX_TEMPLATE_URL: Ansible Tower endpoint URL to launch the appropriate template for this deploy.

Note that the "deploy" stage will be skipped if the earlier build fails, i.e., in our case, if the tests run has failures.